rag_webpage_bot/requirements.txt
flask
openai
faiss-cpu
beautifulsoup4
requests
python-dotenv
numpy
scikit-learn
